class progress 4/3
i added all the logic to get the personal library working it can now remove and edit items in the list and it saves them long term. i also did this for the wishlist minus the editing.

the things i need to do are 1 add a way to update progress of goals. 2 make the goles easer to understand. 3 make yeary and monthly goles avalible. 4 add a way to make percentages avalible to see. 5 i need to be able to edit or remove goals that are in progress.